requirements.txt
tensorflow
pyyaml
opencv-python
matplotlib
dlib
numpy
scikit-image
pillow
torchvision
torch
tqdm
flask
